# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
  description: PBI API Documentation
  termsOfService: http://swagger.io/terms/
  title: PBI API
  version: "1.0"
paths:
...
  /toko:
    get:
      description: Get list of all toko (public)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success get toko list
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get all toko
      tags:
      - Toko
  /toko/{id}:
    get:
      description: Get detail toko by ID
      parameters:
      - description: Toko ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success get toko detail
          schema:
            type: object
        "400":
          description: Invalid toko ID
          schema:
            type: object
        "404":
          description: Toko not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get toko by ID
      tags:
      - Toko
    put:
      consumes:
      - multipart/form-data
      description: Update toko data (owner only)
      parameters:
      - description: Toko ID
        in: path
        name: id
        required: true
        type: integer
      - description: Nama toko
        in: formData
        name: nama_toko
        type: string
      - description: Foto toko
        in: formData
        name: url_foto
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success update toko
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Toko not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update toko
      tags:
      - Toko
  /toko/my:
    get:
      description: Get toko owned by authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Success get my toko
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get my toko
      tags:
      - Toko
  /trx/:
    get:
      consumes:
      - application/json
      description: Get all transactions for authenticated user with full details
      produces:
      - application/json
      responses:
        "200":
          description: Success get all transactions
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get All Transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction with multiple products
      parameters:
      - description: Transaction Request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Transaction created successfully
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Alamat pengiriman tidak valid
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create Transaction
      tags:
      - Transactions
  /trx/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed transaction information by transaction ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success get transaction detail
          schema:
            type: object
        "400":
          description: Invalid transaction ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Transaction not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get Transaction by ID
      tags:
      - Transactions
...
  /user/profile:
    get:
      description: Get authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: Success get profile
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update authenticated user profile
      parameters:
      - description: Update profile payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success update profile
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"